machine_learning/src/requirements.txt
numpy==1.15.4
pandas==0.23.4
mysql-connector==2.1.6